requirements.txt
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
pyarrow==17.0.0
scipy==1.14.1
torch==2.4.0
tqdm==4.66.5
